Symbols to use in colour previews

List the symbols available to use in colour previews.

Usage

list_colour_symbols()

Value

This function is called for its side effects and has no return value.

Details

By default, Unicode characters are used for symbols in colour previews in UTF-8 supported outputs. They automatically fall back to ASCII characters when the output does not support them.

To change the symbol used for colour previews, set the palettes.print_symbol option to a symbol name listed in list_colour_symbols().

See also

Examples

list_colour_symbols()
#> ── options(palettes.print_symbol = ...) ────────────────────
#> •	circle_small            ⬛	square_large           
#> ●	circle                  ▇	block                   
#> ⚫	circle_medium          █	block_full              
#> ⬤	circle_large            ⬩	diamond_small           
#> ⬬	ellipse_horizontal      ◆	diamond                 
#> ⬮	ellipse_vertical        ⬥	diamond_medium          
#> ▲	triangle_up             ⬟	pentagon                
#> ▼	triangle_down           ⬢	hexagon                 
#> ◀	triangle_left           ★	star                    
#> ▶	triangle_right          ♥	heart                   
#> ▪	square_small            ☺	smiley                  
#> ■	square                  ෴	moustache               
#> ◼	square_medium