Colour palette interpolation

Interpolate the set of colours in palettes_palette or palettes_colour objects to create new colour palettes.

Usage

pal_brewer(
  palette,
  n = NULL,
  direction = 1,
  interpolate = c("linear", "spline")
)

Arguments

palette

An object of class palettes_palette or palettes_colour.

n

An integer specifying the number of colours to return.

direction

Sets the order of colours in the scale. If 1, the default, colours are ordered from first to last If -1, the order of colours is reversed.

interpolate

The interpolation method. Either "linear" (default) or "spline".

Value

An object of class palettes_palette or palettes_colour.

Examples

# The class returned after interpolation matches the input class.
x <- pal_colour(c("red", "green", "blue"))
y <- pal_palette(my_pal = x)
pal_brewer(x)
#> <palettes_colour[3]>
#> • #FF0000
#> • #00FF00
#> • #0000FF
pal_brewer(y)
#> <palettes_palette[1]>
#> $my_pal
#> <palettes_colour[3]>
#> • #FF0000
#> • #00FF00
#> • #0000FF
#> 

# Choose between linear and spline interpolation.
pal_brewer(x, n = 7, interpolate = "linear")
#> <palettes_colour[7]>
#> • #FF0000
#> • #DE8A00
#> • #AAC800
#> • #00FF00
#> • #77B784
#> • #746EC4
#> • #0000FF
pal_brewer(x, n = 7, interpolate = "spline")
#> <palettes_colour[7]>
#> • #FF0000
#> • #C9BB00
#> • #7AF400
#> • #00FF00
#> • #23E070
#> • #599AB7
#> • #0000FF

# Palettes will have the same length after interpolation, regardless of the
# number of colours in the original palette.
z <- pal_palette(
  pal1 = c("red", "green", "blue"),
  pal2 = c("orange", "purple")
)
pal_brewer(z, n = 5)
#> <palettes_palette[2]>
#> $pal1
#> <palettes_colour[5]>
#> • #FF0000
#> • #C8AB00
#> • #00FF00
#> • #7C93A6
#> • #0000FF
#> 
#> $pal2
#> <palettes_colour[5]>
#> • #FFA500
#> • #F3875E
#> • #E16891
#> • #C847C0
#> • #9F1FF0
#>